Ejercicio2/requirements.txt
django
djangorestframework
drf-yasg
coreapi
pytest
pytest-django
pytest-drf
ruff